System logger is presented with new abstract interface now. Code that works with system logger can also work with any other logger implementing this interface. So it's easier now to write code for cocaine application that can be used without worker and service manager.
